renamed registration_utilities.h to registration_utilities.hpp
