+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_constitutive/test_constitutive_law_dimensions.cpp
@@ -11,7 +11,7 @@
 //
 #include "custom_constitutive/plane_strain.h"
 #include "custom_constitutive/three_dimensional.h"
-#include "custom_utilities/registration_utilities.h"
+#include "custom_utilities/registration_utilities.hpp"
 #include "includes/constitutive_law.h"
 #include "includes/stream_serializer.h"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
