+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_elements/test_interface_stress_state_policy.cpp
@@ -12,7 +12,7 @@
 
 #include "custom_elements/interface_stress_state.h"
 #include "custom_geometries/interface_geometry.hpp"
-#include "custom_utilities/registration_utilities.h"
+#include "custom_utilities/registration_utilities.hpp"
 #include "includes/stream_serializer.h"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 #include "tests/cpp_tests/test_utilities.h"
