+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_elements/test_pw_element.cpp
@@ -20,7 +20,7 @@
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 #include "tests/cpp_tests/test_utilities.h"
 
-#include "custom_utilities/registration_utilities.h"
+#include "custom_utilities/registration_utilities.hpp"
 #include <boost/numeric/ublas/assignment.hpp>
 
 namespace
